meu-servidor-ia/services/dashboard/requirements.txt
# Framework web e utilitários
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.2
pydantic-settings>=2.1.0
email-validator>=2.1.0.post1

# Banco de dados e migrações
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
alembic>=1.7.1

# Cache
aioredis>=2.0.1

# Templates e frontend
jinja2>=3.1.2
python-dotenv>=1.0.0

# Logging e monitoramento
loguru>=0.7.2

# Testes
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

# Monitoramento
psutil>=5.8.0
aiohttp>=3.8.0

# Segurança
bcrypt>=3.2.0

# Utilitários
python-dateutil>=2.8.2
humanize>=3.11.0 